Move constants.dart to Utils folder
